+++ Team6220_PowerPlay/src/main/java/org/firstinspires/ftc/team6220_PowerPlay/testclasses/AprilTagTest.java
@@ -17,13 +17,7 @@ public class AprilTagTest extends BaseAutonomous {
     @Override
     public void runOpMode() throws InterruptedException {
         initialize();
-
         detectSignal();
-
         waitForStart();
-
-        initializeCameras(Constants.LOWER_RED, Constants.UPPER_RED);
-
-        driveInches(0, 12);
     }
 }
