+++ Team6220_PowerPlay/src/main/java/org/firstinspires/ftc/team6220_PowerPlay/testclasses/CenterJunctionBottom.java
@@ -2,17 +2,41 @@ package org.firstinspires.ftc.team6220_PowerPlay.testclasses;
 
 import com.qualcomm.robotcore.eventloop.opmode.Autonomous;
 
+import org.firstinspires.ftc.robotcore.external.hardware.camera.WebcamName;
 import org.firstinspires.ftc.team6220_PowerPlay.BaseAutonomous;
 import org.firstinspires.ftc.team6220_PowerPlay.Constants;
+import org.firstinspires.ftc.team6220_PowerPlay.RobotCameraPipeline;
+import org.openftc.easyopencv.OpenCvCamera;
+import org.openftc.easyopencv.OpenCvCameraFactory;
+import org.openftc.easyopencv.OpenCvCameraRotation;
 
 //@Disabled
 @Autonomous(name = "CenterJunctionBottom", group = "Test")
 public class CenterJunctionBottom extends BaseAutonomous {
 
+    public RobotCameraPipeline robotCameraPipeline;
+    OpenCvCamera camera;
+
     @Override
     public void runOpMode() throws InterruptedException {
         initialize();
-        initializeCameras(Constants.LOWER_YELLOW, Constants.UPPER_YELLOW);
+
+        int cameraMonitorViewId = hardwareMap.appContext.getResources().getIdentifier("cameraMonitorViewId", "id", hardwareMap.appContext.getPackageName());
+        camera = OpenCvCameraFactory.getInstance().createWebcam(hardwareMap.get(WebcamName.class, "RobotCamera"), cameraMonitorViewId);
+
+        robotCameraPipeline = new RobotCameraPipeline();
+        robotCameraPipeline.setRanges(Constants.LOWER_YELLOW, Constants.LOWER_YELLOW);
+
+        camera.setPipeline(robotCameraPipeline);
+        camera.openCameraDeviceAsync(new OpenCvCamera.AsyncCameraOpenListener() {
+            @Override
+            public void onOpened() {
+                camera.startStreaming(Constants.CAMERA_X, Constants.CAMERA_Y, OpenCvCameraRotation.UPRIGHT);
+            }
+
+            @Override
+            public void onError(int errorCode) {}
+        });
 
         waitForStart();
